PoC Final/requirements.txt
pandas==1.3.4
matplotlib==3.1.3
nltk==3.6.5
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
textdistance==4.2.1
ngram==4.0.3
pickle5==0.0.12
Unidecode==1.3.3
python-Levenshtein==0.12.2
fuzzywuzzy==0.18.0
pandarallel==1.5.7
langdetect==1.0.9
pycld3==0.22
numpydoc
sphinx
sphinxcontrib-napoleon

